revert to jar packaging, cleanup tester unpack, remove overlay
