+++ open-health-manager-app/pom.xml
@@ -12,7 +12,7 @@
     </parent>
 
     <artifactId>open-health-manager-app</artifactId>
-    <packaging>war</packaging>
+    <packaging>jar</packaging>
     <name>Open Health Manager App</name>
     <description>Open Health Manager App</description>
 
@@ -129,75 +129,6 @@
             <type>war</type>
             <scope>provided</scope>
         </dependency>
-        <dependency>
-            <groupId>ca.uhn.hapi.fhir</groupId>
-            <artifactId>hapi-fhir-testpage-overlay</artifactId>
-            <version>${hapi.fhir.version}</version>
-            <classifier>classes</classifier>
-	        </dependency>
-		<!-- webjars -->
-		<dependency>
-			<groupId>org.webjars.npm</groupId>
-			<artifactId>bootstrap</artifactId>
-			<version>4.5.2</version>
-		</dependency>
-		<dependency>
-			<groupId>org.webjars</groupId>
-			<artifactId>Eonasdan-bootstrap-datetimepicker</artifactId>
-			<version>4.17.47</version>
-			<exclusions>
-				<exclusion>
-					<groupId>*</groupId>
-					<artifactId>*</artifactId>
-				</exclusion>
-			</exclusions>
-		</dependency>
-		<dependency>
-			<groupId>org.webjars</groupId>
-			<artifactId>font-awesome</artifactId>
-			<version>5.8.2</version>
-		</dependency>
-		<dependency>
-			<groupId>org.webjars.bower</groupId>
-			<artifactId>awesome-bootstrap-checkbox</artifactId>
-			<version>1.0.2</version>
-		</dependency>
-		<dependency>
-			<groupId>org.webjars</groupId>
-			<artifactId>jstimezonedetect</artifactId>
-			<version>1.0.6</version>
-		</dependency>
-		<dependency>
-			<groupId>org.webjars</groupId>
-			<artifactId>select2</artifactId>
-			<version>4.0.13</version>
-			<exclusions>
-				<exclusion>
-					<groupId>org.webjars</groupId>
-					<artifactId>jquery</artifactId>
-				</exclusion>
-			</exclusions>
-		</dependency>
-		<dependency>
-			<groupId>org.webjars.bower</groupId>
-			<artifactId>jquery</artifactId>
-			<version>3.5.1</version>
-		</dependency>
-		<dependency>
-			<groupId>org.webjars.bower</groupId>
-			<artifactId>moment</artifactId>
-			<version>2.27.0</version>
-		</dependency>
-		<dependency>
-			<groupId>org.webjars.npm</groupId>
-			<artifactId>popper.js</artifactId>
-			<version>1.16.1</version>
-		</dependency>
-		<dependency>
-			<groupId>org.webjars</groupId>
-			<artifactId>swagger-ui</artifactId>
-			<version>4.1.3</version>
-		</dependency>
 		<!-- JHipster generated -->          
         <dependency>
             <groupId>tech.jhipster</groupId>
@@ -471,26 +402,46 @@
                 <groupId>org.gaul</groupId>
                 <artifactId>modernizer-maven-plugin</artifactId>
             </plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-dependency-plugin</artifactId>
+				<executions>
+					<execution>
+						<id>unpacktester</id>
+						<phase>compile</phase>
+						<goals>
+							<goal>unpack</goal>
+						</goals>
+						<configuration>
+							<artifactItems>
+								<artifactItem>
+									<groupId>ca.uhn.hapi.fhir</groupId>
+									<artifactId>hapi-fhir-testpage-overlay</artifactId>
+									<version>${hapi.fhir.version}</version>
+									<type>war</type>
+									<overWrite>true</overWrite>
+									<outputDirectory>${project.build.directory}/classes/static
+									</outputDirectory>
+									<includes>**/*.js,**/*.css,**/img/**</includes>
+								</artifactItem>
+								<artifactItem>
+									<groupId>ca.uhn.hapi.fhir</groupId>
+									<artifactId>hapi-fhir-testpage-overlay</artifactId>
+									<version>${hapi.fhir.version}</version>
+									<type>war</type>
+									<overWrite>true</overWrite>
+									<outputDirectory>${project.build.directory}/classes/tester
+									</outputDirectory>
+									<includes>**/WEB-INF/templates/**</includes>
+								</artifactItem>								
+							</artifactItems>
+							<overWriteReleases>false</overWriteReleases>
+							<overWriteSnapshots>true</overWriteSnapshots>
+						</configuration>
+					</execution>
+				</executions>
+			</plugin>             
             <!-- jhipster-needle-maven-add-plugin -->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-war-plugin</artifactId>
-                <configuration>
-                    <archive>
-                        <manifestEntries>
-                            <Build-Time>${maven.build.timestamp}</Build-Time>
-                        </manifestEntries>
-                    </archive>
-                    <attachClasses>true</attachClasses>
-                    <overlays>
-                        <overlay>
-                            <groupId>ca.uhn.hapi.fhir</groupId>
-                            <artifactId>hapi-fhir-testpage-overlay</artifactId>
-                        </overlay>
-                    </overlays>
-                    <failOnMissingWebXml>false</failOnMissingWebXml>
-                </configuration>
-            </plugin>              
         </plugins>
         <pluginManagement>
             <plugins>
@@ -711,7 +662,6 @@
                             <image>openhealthmanagerapp:latest</image>
                         </to>
                         <container>
-                            <appRoot>/app</appRoot>
                             <entrypoint>
                                 <shell>bash</shell>
                                 <option>-c</option>
@@ -1089,35 +1039,7 @@
                     <plugin>
                         <groupId>org.liquibase</groupId>
                         <artifactId>liquibase-maven-plugin</artifactId>
-                    </plugin>
-			  		<plugin>
-			  			<groupId>org.apache.maven.plugins</groupId>
-			  			<artifactId>maven-dependency-plugin</artifactId>
-			  			<executions>
-			  				<execution>
-			  					<id>unpack</id>
-			  					<phase>compile</phase>
-			  					<goals>
-			  						<goal>unpack</goal>
-			  					</goals>
-			  					<configuration>
-			  						<artifactItems>
-			  							<artifactItem>
-								            <groupId>ca.uhn.hapi.fhir</groupId>
-								            <artifactId>hapi-fhir-testpage-overlay</artifactId>
-								            <version>${hapi.fhir.version}</version>
-								            <type>war</type>
-			  								<overWrite>false</overWrite>
-			  								<outputDirectory>${project.build.directory}/classes/static</outputDirectory>
-			  								<includes>**/*.js,**/*.css,**/img/**</includes>
-			  							</artifactItem>
-			  						</artifactItems>
-			  						<overWriteReleases>false</overWriteReleases>
-			  						<overWriteSnapshots>true</overWriteSnapshots>
-			  					</configuration>
-			  				</execution>
-			  			</executions>
-			  		</plugin>                    
+                    </plugin>                 
                 </plugins>
                 <pluginManagement>
                     <plugins>
