+++ open-health-manager-fhir/src/main/java/org/mitre/healthmanager/fhir/config/TesterOverlayForwardingController.java
@@ -1,8 +1,8 @@
 package org.mitre.healthmanager.fhir.config;
 
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
 /**
@@ -10,71 +10,47 @@ import org.springframework.web.bind.annotation.PostMapping;
  * Will override other mappings.
  */
 @Controller
+@ConditionalOnExpression("'${hapi.fhir.tester}' != null")
 public class TesterOverlayForwardingController {
     @GetMapping("/home")
     public String forwardHome() {
-        return "forward:/tester/home";
+        return "redirect:/tester/home";
     }
     @GetMapping("/about")
     public String forwardAbout() {
-        return "forward:/tester/about";
+        return "redirect:/tester/about";
     }
     
     @GetMapping("/resource")
     public String forwardResource() {
-        return "forward:/tester/resource";
+        return "redirect:/tester/resource";
     }
     @GetMapping("/search")
     public String forwardSearch() {
-        return "forward:/tester/search";
+        return "redirect:/tester/search";
     }
     @GetMapping("/read")
     public String forwardRead() {
-        return "forward:/tester/read";
+        return "redirect:/tester/read";
     }
     @GetMapping("/history-type")
     public String forwardHistoryType() {
-        return "forward:/tester/history-type";
+        return "redirect:/tester/history-type";
     }
     @GetMapping("/delete")
     public String forwardDelete() {
-        return "forward:/tester/delete";
+        return "redirect:/tester/delete";
     }
     @PostMapping("/create")
     public String forwardCreate() {
-        return "forward:/tester/create";
+        return "redirect:/tester/create";
     }
     @PostMapping("/update")
     public String forwardUpdate() {
-        return "forward:/tester/update";
+        return "redirect:/tester/update";
     }
     @PostMapping("/validate")
     public String forwardValidate() {
-        return "forward:/tester/validate";
-    }
-    
-    @GetMapping(path = {"/css/{res}"})
-    public String forwardCss(@PathVariable String res) {
-        return "forward:/tester/css/" + res;
-    }
-
-    @GetMapping(path = {"/img/{res}"})
-    public String forwardImg(@PathVariable String res) {
-        return "forward:/tester/img/" + res;
-    }
-    
-    @GetMapping(path = {"/js/{res}"})
-    public String forwardJs(@PathVariable String res) {
-        return "forward:/tester/js/" + res;
-    }
-    
-    @GetMapping(path = {"/fa/{res}"})
-    public String forwardFa(@PathVariable String res) {
-        return "forward:/tester/fa/" + res;
-    }
-    
-    @GetMapping(path = {"/fonts/{res}"})
-    public String forwardFonts(@PathVariable String res) {
-        return "forward:/tester/fonts/" + res;
+    	return "redirect:/tester/validate";
     }
 }
